+++ NEWS.md
@@ -1,5 +1,7 @@
 # newpost 0.1.0
 
+* Added see_cats() function to view available category on a standard Quart blog home page.
+
 # newpost 0.0.0.9000
 
 * Added main newpost() function to create a folder and index.qmd file with YAML for a new Quarto blog post.